Split memory functions and structs in pal into separate header
